requirements.txt
# Web 框架
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.4
Werkzeug==2.3.7

# 数据库
PyMySQL==1.1.0
SQLAlchemy==2.0.20
redis==5.0.0

# 安全认证
bcrypt==4.0.1
PyJWT==2.8.0
email-validator==2.0.0

# 文件处理
openpyxl==3.1.2
ReportLab==4.0.4

# 数据分析
pandas==2.0.3
scikit-learn==1.3.0

# HTTP 请求
requests==2.31.0

# 部署
gunicorn==21.2.0

# 监控
sentry-sdk[flask]==1.31.0

# 工具
python-dotenv==1.0.0
python-dateutil==2.8.2

# 测试
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.3.0
